Portfolio module fix storage upgrade
