+++ pallets/asset/src/lib.rs
@@ -411,13 +411,13 @@ decl_module! {
             storage_migrate_on!(storage_ver, 1, {
                 let mut id_map = BTreeMap::<_, u32>::new();
                 migrate_double_map::<_, _, Blake2_128Concat, _, _, _, _, _>(
-                b"Asset", b"AssetDocuments",
+                    b"Asset", b"AssetDocuments",
                     |ticker: Ticker, name: DocumentName, doc: DocumentOld| {
                         let count = id_map.entry(ticker).or_default();
                         let id = DocumentId(mem::replace(count, *count + 1));
                         Some((ticker, id, doc.migrate(name)?))
                     }
-                    );
+                );
                 for (ticker, id) in id_map {
                     AssetDocumentsIdSequence::insert(ticker, DocumentId(id));
                 }
