+++ pallets/portfolio/src/lib.rs
@@ -41,26 +41,31 @@
 //! - `ensure_portfolio_custody`: Makes sure that the given identity has custodian access over the portfolio.
 //! - `ensure_portfolio_transfer_validity`: Makes sure that a transfer between two portfolios is valid.
 
+#![feature(const_option)]
+
 #![cfg_attr(not(feature = "std"), no_std)]
 
 #[cfg(feature = "runtime-benchmarks")]
+
 pub mod benchmarking;
 
 use codec::{Decode, Encode};
 use frame_support::{
     decl_error, decl_event, decl_module, decl_storage, dispatch::DispatchResult, ensure,
-    weights::Weight, IterableStorageDoubleMap,
+    storage::StorageValue, weights::Weight, IterableStorageDoubleMap,
 };
 use pallet_identity::{self as identity, PermissionedCallOriginData};
 use polymesh_common_utilities::{
     identity::Trait as IdentityTrait, traits::portfolio::PortfolioSubTrait, CommonTrait,
 };
 use polymesh_primitives::{
-    AuthorizationData, AuthorizationError, IdentityId, PortfolioId, PortfolioKind, PortfolioName,
-    PortfolioNumber, SecondaryKey, Signatory, Ticker,
+    storage_migration_ver, AuthorizationData, AuthorizationError, IdentityId, PortfolioId,
+    PortfolioKind, PortfolioName, PortfolioNumber, SecondaryKey, Signatory, Ticker,
 };
 use sp_arithmetic::traits::{CheckedSub, Saturating};
 use sp_std::{iter, mem, prelude::Vec};
+
+storage_migration_ver!(1);
 
 type Identity<T> = identity::Module<T>;
 
@@ -113,6 +118,8 @@ decl_storage! {
         /// `false` values are never explicitly stored in the map, and are instead inferred by the absence of a key.
         pub PortfoliosInCustody get(fn portfolios_in_custody):
             double_map hasher(identity) IdentityId, hasher(twox_64_concat) PortfolioId => bool;
+        /// Storage version.
+        StorageVersion get(fn storage_version) build(|_| Version::new(1).unwrap()): Version;
     }
 }
 
@@ -202,6 +209,23 @@ decl_module! {
 
         /// The event logger.
         fn deposit_event() = default;
+
+        fn on_runtime_upgrade() -> Weight {
+            use polymesh_primitives::{storage_migrate_on, migrate::copy_map_rename_module};
+
+            let storage_ver = StorageVersion::get();
+
+            storage_migrate_on!(storage_ver, 1, {
+                copy_map_rename_module::<PortfolioName>(b"Session", b"Portfolio", b"Portfolios");
+                copy_map_rename_module::<T::Balance>(b"Session", b"Portfolio", b"PortfolioAssetBalances");
+                copy_map_rename_module::<PortfolioNumber>(b"Session", b"Portfolio", b"NextPortfolioNumber");
+                copy_map_rename_module::<Option<IdentityId>>(b"Session", b"Portfolio", b"PortfolioCustodian");
+                copy_map_rename_module::<T::Balance>(b"Session", b"Portfolio", b"PortfolioLockedAssets");
+                copy_map_rename_module::<bool>(b"Session", b"Portfolio", b"PortfoliosInCustody");
+            });
+
+            0
+        }
 
         /// Creates a portfolio with the given `name`.
         #[weight = <T as Trait>::WeightInfo::create_portfolio(name.len() as u32)]
