+++ primitives/src/migrate.rs
@@ -70,19 +70,6 @@ pub fn migrate_map<T: Migrate, C: FnMut(&[u8]) -> T::Context>(
     migrate_map_rename::<T, C>(module, module, item, item, derive_context)
 }
 
-/// Migrate the values with old type `T` in `module::item` to `T::Into` and renames the module to
-/// `new_module`.
-///
-/// Migrations mapping to `None` are silently dropped from storage.
-pub fn migrate_map_rename_module<T: Migrate, C: FnMut(&[u8]) -> T::Context>(
-    module: &[u8],
-    new_module: &[u8],
-    item: &[u8],
-    derive_context: C,
-) {
-    migrate_map_rename::<T, C>(module, new_module, item, item, derive_context)
-}
-
 /// Migrate the values with old type `T` in `module::item` to `T::Into` in `module::new_item`.
 ///
 /// Migrations resulting in `old.migrate() == None` are silently dropped from storage.
@@ -187,4 +174,19 @@ pub fn kill_item(module: &[u8], item: &[u8]) {
     prefix[0..16].copy_from_slice(&Twox128::hash(module));
     prefix[16..32].copy_from_slice(&Twox128::hash(item));
     kill_prefix(&prefix)
+}
+
+/// Copies a single or double map storage item under a new module prefix and removes the map from
+/// the old module prefix.
+///
+/// Migrations mapping to `None` are silently dropped from storage.
+pub fn copy_map_rename_module<T: Decode + Encode>(
+    old_module: &[u8],
+    new_module: &[u8],
+    item: &[u8],
+) {
+    StorageIterator::<T>::new(old_module, item)
+        .drain()
+        .filter_map(|(key, val)| Some((key, val)))
+        .for_each(|(key, val)| put_storage_value(new_module, item, &key, val));
 }
